Codewars module
