+++ utils/entities/author.js
@@ -1,5 +1,6 @@
 import fetch from 'node-fetch';
 import { CodewarsKataArray } from './codewarsKata.js';
+import Codewars from './../codewars.js';
 
 class Author {
   constructor(name) {
@@ -7,9 +8,7 @@ class Author {
   }
 
   async initKatas() {
-    const response = await fetch(
-      `https://www.codewars.com/api/v1/users/${this.name}/code-challenges/authored`
-    );
+    const response = await fetch(Codewars.getAuthorsKatasAPILink(this.name));
     if (response.status === 404) {
       this.katas = null;
       return;
